services/weknora/docreader/pyproject.toml
[project]
name = "docreader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.18"
dependencies = [
    "antiword>=0.1.0",
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.14.2",
    "cos-python-sdk-v5>=1.9.38",
    "goose3[all]>=3.1.20",
    "grpcio>=1.76.0",
    "grpcio-health-checking>=1.76.0",
    "grpcio-tools>=1.76.0",
    "lxml>=6.0.2",
    "markdown>=3.10",
    "markdownify>=1.2.0",
    "markitdown[docx,pdf,xls,xlsx]>=0.1.3",
    "minio>=7.2.18",
    "mistletoe>=1.5.0",
    "ollama>=0.6.0",
    "openai>=2.7.1",
    "paddleocr>=2.10.0,<3.0.0",
    "paddlepaddle>=3.0.0,<4.0.0",
    "pdfplumber>=0.11.7",
    "pillow>=12.0.0",
    "playwright>=1.55.0",
    "protobuf>=6.33.0",
    "psutil>=5.9.8",
    "pydantic>=2.12.3",
    "pymupdf>=1.24.0",
    "pypdf>=6.1.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
    "textract==1.5.0",
    "trafilatura>=2.0.0",
    "urllib3>=2.5.0",
]
